Add RangeField to Solr
Index format is in the form [0 TO 10, 7, MIN TO 100, -4 TO MAX], taking up to
four dimensions.

Querying should use the {!pr} query parser, and uses the same format as the
index.  The parser has two required parameters, 'field' specifying the field
to search in, and 'method', which should be 'within', 'contains' or
'intersects'.